+++ eth_defi/erc_4626/core.py
@@ -271,6 +271,11 @@ class ERC4626Feature(enum.Enum):
     #: https://github.com/stratosphere-network/LiquidRoyaltyContracts
     liquidity_royalty_like = "liquidity_royalty_like"
 
+    #: Liquid Royalty
+    #:
+    #: https://www.liquidroyalty.com/vaults
+    liquid_royalty_like = "liquid_royalty_like"
+
     #: cSigma Finance
     #:
     #: https://csigma.finance
@@ -715,6 +720,9 @@ def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
 
     elif ERC4626Feature.liquidity_royalty_like in features:
         return "Liquidity Royalty Tranching"
+
+    elif ERC4626Feature.liquid_royalty_like in features:
+        return "Liquid Royalty"
 
     elif ERC4626Feature.csigma_like in features:
         return "cSigma Finance"